requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Pillow==11.2.1
gunicorn==23.0.0
SQLAlchemy==2.0.41
Werkzeug==3.1.3